Context API Updated to Add Transaction
